square and colours
I created a floor for the player to move the square around on, and also changed the colours of the canvas background so that the floor is more prominent. I also made a square be drawn and set an asset in the square array for when the player can make it jump.