radius size, and testing integral circle
changed the radius variable to be relative to the size of the square. Also i am going to test if i can change it from being a radius to being the shape of the square, i will do this by using the x, y and width/height of the square.